관리자 기부 게시글 승인 시, 기부 회차 상태도 변경

+++ src/main/java/mapper/DonateMapper.java
@@ -24,7 +24,7 @@ public interface DonateMapper {
 	void deleteAction(Long dano);
 	void updateAction(DonateAction action);
 	DonateAction selectOneAction(Long dano);
-	int findMyAmount(@Param("drno") Long drno,@Param("mno") Long mno);
+	long findMyAmount(@Param("drno") Long drno,@Param("mno") Long mno);
 	long findMyTotalAmount(Long mno);
 	long findTotalAmount();
 }
